[hotfix][iteration] Merge the IterationFactory and Iterations
